requirements.txt
numpy~=1.24.2
scipy~=1.10.0
requests~=2.28.2
opencv-python~=4.7.0.68
matplotlib~=3.7.0
torch~=1.13.1
Pillow~=9.4.0
torchvision~=0.14.1
pypdfium2~=3.21.0
weasyprint~=58.0
defusedxml~=0.7.1
tqdm~=4.64.1
Unidecode~=1.3.6
mplcursors~=0.5.2
langdetect~=1.0.9
pyclipper~=1.3.0.post4
Shapely~=1.8.5.post1
rapidfuzz~=2.13.7
h5py~=3.8.0
setuptools~=66.1.1
psutil~=5.9.4
pandas~=1.5.3
protobuf==3.20.1
pytesseract
transformers
sentencepiece
pybind11
fastwer
jupyter
wandb
fastprogress